Added a safe_ltf8_decode function to go along with the itf8 variant.
Switched the record_counter from 32-bit to 64-bit.  This is as per the
spec, but it was an error that we only stored it in 32-bit.

(Plus a couple tiny minor tweaks to spacing, to make diffing to htslib
easier.)

+++ io_lib/cram_structs.h
@@ -326,7 +326,7 @@ typedef struct {
     int32_t ref_seq_start;  /* if content_type == MAPPED_SLICE */
     int32_t ref_seq_span;   /* if content_type == MAPPED_SLICE */
     int32_t num_records;
-    int32_t record_counter;
+    int64_t record_counter;
     int32_t num_blocks;
     int32_t num_content_ids;
     int32_t *block_content_ids;
@@ -353,7 +353,7 @@ typedef struct {
     int32_t  ref_seq_id;
     int32_t  ref_seq_start;
     int32_t  ref_seq_span;
-    int32_t  record_counter;
+    int64_t  record_counter;
     int64_t  num_bases;
     int32_t  num_records;
     int32_t  num_blocks;
